+++ CHANGELOG.md
@@ -13,6 +13,10 @@ Changelog
  - Button classes are insert at a tag for hyperlink elements (See #22)
  - Add responsive-embed classes for youtube and vimeo player having aspect ratio ([#21](https://github.com/contao-bootstrap/templates/issues/21))
  - Backport playerAspect for video players so that responsive-embed is usable in Contao 4.4 ([#21](https://github.com/contao-bootstrap/templates/issues/21))
+ 
+### Changed
+
+ - Use `grid.columns` configuration if available for grid gallery template (#20)
 
 2.0.0 (2018-08-01)
 ------------------